requirements.txt
anthropic
python-dotenv
requests
sendgrid
slack-sdk